add changes in the styling component SignIn
