+++ frontend/src/Components/Buttons.css
@@ -5,7 +5,8 @@
   border: none;
   border-radius: 6px;
   cursor: pointer;
-  margin:3px;
+  margin: 3px 0;
+  width: 100%;
 }
 
 .button-container button:hover {
