+++ frontend/src/Components/Header.css
@@ -15,7 +15,7 @@
 
 .header-container h1 {
   text-align: center;
-  font-size: 2.5rem;
+  font-size: 2rem;
   backdrop-filter: blur(2px);
   margin: 20px;
   padding: 0px 15px;
