+++ frontend/src/Components/SignInForm.css
@@ -1,0 +1,46 @@
+.sign-in-container {
+  z-index: 10;
+  position: absolute;
+  background-color: rgba(255, 255, 255, 0.837);
+  box-shadow: 2px 2px 20px 10px rgba(0, 0, 0, 0.15);
+  display: flex;
+  flex-direction: column;
+  justify-content: space-between;
+  align-items: center;
+  padding: 5px;
+  width: 100%;
+  height: 170px;
+  backdrop-filter: blur(2px);
+  border-radius: 6px;
+  top: 0;
+  left: 0;
+}
+
+.input-field {
+  padding: 0 5px;
+  margin: 5px;
+  display: flex;
+  border-radius: 6px;
+  background-color: white;
+  border: 2px solid rgb(162, 162, 162);
+  flex-direction: row;
+  align-items: center;
+  justify-content: space-between;
+  width: 100%;
+}
+
+.sign-in-container input {
+  padding: 8px 16px;
+  padding-left: 1px;
+  border: 0;
+  cursor: pointer;
+  width: 100%;
+  margin: 3px 0;
+}
+
+
+/* TABLET SCREENS min 745px ------------------------------------------------------------ */
+
+@media (min-width: 745px) {
+
+}
